+++ alembic/versions/db3c91261770_add_temporal_search_attributes.py
@@ -1,0 +1,61 @@
+"""Add Temporal search attributes
+
+Revision ID: db3c91261770
+Revises: 849db8d3b59d
+Create Date: 2025-01-18 21:27:38.967454
+
+"""
+
+import asyncio
+from collections.abc import Sequence
+
+from temporalio.api.enums.v1 import IndexedValueType
+from temporalio.api.operatorservice.v1 import (
+    AddSearchAttributesRequest,
+    RemoveSearchAttributesRequest,
+)
+
+from tracecat.config import TEMPORAL__CLUSTER_NAMESPACE
+from tracecat.dsl.client import get_temporal_client
+from tracecat.logger import logger
+
+# revision identifiers, used by Alembic.
+revision: str = "db3c91261770"
+down_revision: str | None = "849db8d3b59d"
+branch_labels: str | Sequence[str] | None = None
+depends_on: str | Sequence[str] | None = None
+
+
+def upgrade() -> None:
+    async def add_temporal_search_attributes():
+        client = await get_temporal_client()
+        await client.operator_service.add_search_attributes(
+            AddSearchAttributesRequest(
+                search_attributes={
+                    "TracecatTriggerType": IndexedValueType.INDEXED_VALUE_TYPE_KEYWORD,
+                    "TracecatTriggeredByUserId": IndexedValueType.INDEXED_VALUE_TYPE_KEYWORD,
+                },
+                namespace=TEMPORAL__CLUSTER_NAMESPACE,
+            )
+        )
+        logger.info("Temporal search attributes added")
+
+    # ### commands auto generated by Alembic - please adjust! ###
+    asyncio.run(add_temporal_search_attributes())
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    async def remove_temporal_search_attributes():
+        client = await get_temporal_client()
+        await client.operator_service.remove_search_attributes(
+            RemoveSearchAttributesRequest(
+                search_attributes=["TracecatTriggerType", "TracecatTriggeredByUserId"],
+                namespace=TEMPORAL__CLUSTER_NAMESPACE,
+            )
+        )
+        logger.info("Temporal search attributes removed")
+
+    # ### commands auto generated by Alembic - please adjust! ###
+    asyncio.run(remove_temporal_search_attributes())
+    # ### end Alembic commands ###
